The satellite now revolves around the earth

+++ earth.hpp
@@ -4,7 +4,7 @@
 namespace Simulation {
     struct Earth {
         static int Radius, RadiusOriginal;
-        static int X, Y;
+        static double X, Y;
 
         static void Initialize(const ID2D1Bitmap* const bmp);
     };
